+++ src/main/java/com/aetherteam/aether/command/PlayerAttachmentCommand.java
@@ -1,6 +1,7 @@
 package com.aetherteam.aether.command;
 
 import com.aetherteam.aether.attachment.AetherDataAttachments;
+import com.aetherteam.aether.attachment.AetherPlayerAttachment;
 import com.aetherteam.aether.network.packet.clientbound.HealthResetPacket;
 import com.aetherteam.aether.attachment.AttachmentSyncable;
 import com.mojang.brigadier.CommandDispatcher;
@@ -54,7 +55,7 @@ public class PlayerAttachmentCommand {
             ServerPlayer player = playerList.getPlayer(gameProfile.id());
             if (player != null) {
                 var data = player.getAttachedOrCreate(AetherDataAttachments.AETHER_PLAYER);
-                data.setSynced(player.getId(), AttachmentSyncable.Direction.CLIENT, "setLifeShardCount", value);
+                data.setSynced(player.getId(), AttachmentSyncable.Direction.CLIENT, AetherPlayerAttachment.LIFE_SHARD_COUNT_SYNC_KEY, value);
                 AttributeInstance attribute = player.getAttribute(Attributes.MAX_HEALTH);
                 if (attribute != null) {
                     attribute.removeModifier(data.getLifeShardHealthAttributeModifier().id());