+++ src/main/java/com/aetherteam/aether/item/miscellaneous/LifeShardItem.java
@@ -1,6 +1,7 @@
 package com.aetherteam.aether.item.miscellaneous;
 
 import com.aetherteam.aether.attachment.AetherDataAttachments;
+import com.aetherteam.aether.attachment.AetherPlayerAttachment;
 import com.aetherteam.aether.attachment.AttachmentSyncable;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
@@ -32,7 +33,7 @@ public class LifeShardItem extends Item implements ConsumableItem {
                 player.swing(hand);
                 if (!level.isClientSide()) {
                     this.consume(this, heldStack, player);
-                    aetherPlayer.setSynced(player.getId(), AttachmentSyncable.Direction.CLIENT, "setLifeShardCount", aetherPlayer.getLifeShardCount() + 1);
+                    aetherPlayer.setSynced(player.getId(), AttachmentSyncable.Direction.CLIENT, AetherPlayerAttachment.LIFE_SHARD_COUNT_SYNC_KEY, aetherPlayer.getLifeShardCount() + 1);
                     return InteractionResult.CONSUME;
                 } else {
                     return InteractionResult.SUCCESS;